requirements.txt
python-dotenv
langchain-core
langchain-community
langchain-google-genai
langchain
